+++ internal/dto/request/get_group_info.go
@@ -1,0 +1,5 @@
+package request
+
+type GetGroupInfoRequest struct {
+	GroupID int `json:"group_id"`
+}