+++ internal/http_server/chat.go
@@ -12,4 +12,7 @@ func init() {
 	GE = gin.Default()
 	GE.POST("/register", v1.Register)
 	GE.POST("/login", v1.Login)
+	GE.GET("/user/info", v1.GetUserInfo)
+	GE.POST("/group/create", v1.CreateGroup)
+	GE.GET("/group/info", v1.GetGroupInfo)
 }